+++ packages/components/src/index.ts
@@ -13,7 +13,7 @@ export * from './components/buttons/Button/Button';
 export * from './components/buttons/PinButton/PinButton';
 export * from './components/buttons/TooltipButton/TooltipButton';
 export * from './components/Flag/Flag';
-export * from './components/Dropdown';
+export * from './components/Dropdown/Dropdown';
 export * from './components/form/Input';
 export * from './components/form/Textarea';
 export * from './components/form/Select';
